create new utils folder and move db utils into it
